requirements.txt
PyYAML==6.0
pytest==7.1.2
